+++ src/plat/normal/load.rs
@@ -5,7 +5,7 @@ use crate::plat::interfaces::load::LoadInterface;
 use super::{cpu_plat::CpuPlat, mesh::Mesh};
 
 impl LoadInterface for CpuPlat {
-    fn load_chunk(&self, position: glam::UVec3, lod_level: u8) -> Box<Chunk> {
+    fn load_chunk(&self, position: glam::UVec3, lod_level: usize) -> Box<Chunk> {
         Box::new(
             self.borrow_raw_plat()
                 .load_chunk(position.to_array().into(), lod_level),
